publish-fq.sh - v1.1 - Minor bug correction
When Sample_ID and Sample_Name in the sample sheet are different, bcl2fastq creates directory named after the Sample_ID and fastq file named after the Sample_Name. Because the script was looking for fastq file named after the Sample_ID, it was not able to find the expected file in this particular case ending with a no file found error.